# openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: RESTful API for managing todo items with soft delete support
  contact: {}
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://todo.local
    description: Kubernetes (local)
security: []
paths:
  /api/v1/todos:
    get:
      operationId: listTodos
      description: Returns all todos, optionally filtered.
      parameters:
        - description: Filter by completion status
          in: query
          name: completed
          schema:
            type: boolean
        - description: Case-insensitive title search
          in: query
          name: search
          schema:
            type: string
        - description: Due before (RFC3339)
          in: query
          name: due_before
          schema:
            type: string
        - description: Due after (RFC3339)
          in: query
          name: due_after
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/model.Todo"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: List todos
      tags:
        - todos
    post:
      operationId: createTodo
      description: Creates a new todo from the request body.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handler.createTodoRequest"
        description: Todo request body
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.Todo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Create a todo
      tags:
        - todos
  "/api/v1/todos/{id}":
    delete:
      operationId: deleteTodo
      description: Deletes a todo by ID. This is a soft delete (via gorm.DeletedAt) —
        the row stays in the database with a deleted_at timestamp.
      parameters:
        - description: Todo ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Delete a todo (soft delete)
      tags:
        - todos
    get:
      operationId: getTodo
      description: Returns a todo by its identifier.
      parameters:
        - description: Todo ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.Todo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Get a todo by ID
      tags:
        - todos
    put:
      operationId: updateTodo
      description: Updates a todo's title, completion status, due date, or tags.
      parameters:
        - description: Todo ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handler.updateTodoRequest"
        description: Todo update request body
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.Todo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Update a todo
      tags:
        - todos
  /api/v1/todos/clear-completed:
    post:
      operationId: clearCompleted
      description: Deletes every todo whose completed flag is true. This is a soft delete
        (via gorm.DeletedAt).
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Delete all completed todos (soft delete)
      tags:
        - todos
  /api/v1/todos/deleted:
    get:
      operationId: listDeletedTodos
      description: Returns every soft-deleted todo.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/model.Todo"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: List deleted todos
      tags:
        - todos
  /api/healthz:
    get:
      operationId: getHealthz
      description: Returns 200 OK as long as the service is running. Exposed externally via ingress at /api/healthz.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Liveness check
      tags:
        - health
  /api/readyz:
    get:
      operationId: getReadyz
      description: Returns 200 OK once the service is ready to accept traffic (database
        connection is healthy). Exposed externally via ingress at /api/readyz.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Readiness check
      tags:
        - health
components:
  schemas:
    handler.createTodoRequest:
      properties:
        due_date:
          type: string
        tags:
          items:
            type: string
          type: array
        title:
          maxLength: 255
          type: string
      required:
        - title
      type: object
    handler.updateTodoRequest:
      properties:
        completed:
          type: boolean
        due_date:
          type: string
        tags:
          items:
            type: string
          type: array
        title:
          maxLength: 255
          type: string
      required:
        - title
      type: object
    model.Todo:
      properties:
        completed:
          description: Completion status
          type: boolean
        created_at:
          description: Creation timestamp
          format: date-time
          type: string
        deleted_at:
          description: Soft-delete timestamp
          format: date-time
          type: string
        due_date:
          description: Due date
          example: 2023-12-31T23:59:59Z
          format: date-time
          type: string
        id:
          description: Unique identifier
          type: integer
        tags:
          description: Todo tags
          items:
            type: string
          type: array
        title:
          description: Todo title (max 255 characters)
          type: string
        updated_at:
          description: Last-updated timestamp
          format: date-time
          type: string
      type: object